Provide an abstract API for a stream that one can write bytes into.
For QUIC, we have a QuicPacketWriter as an API for outputting all datagrams used by the QUIC connection; since encapsulated WebTransport (aka WebTransport over HTTP/2) has a single bytestream as an output, we need a similar type for that.

This CL takes the existing WebTransport stream write API, cleans it up, and moves it into its own class.

Notable changes to the write API:
- Make it vectorized
- Let it return a status object instead of simple true/false
- Merge writing and sending fin into a single method, the same way as existing QUIC APIs work

Parts of the old API are still kept around for compatibility with existing calls.

PiperOrigin-RevId: 501849717