Update QuicWriteBlockedList tests and comments.
Add kStatic and kNonStatic for expressiveness and also for consistency with
kNonIncremental (and future kIncremental).

Make QuicWriteBlockedListTest not inherit from QuicTestWithParam<>, as the tests
are not actually parametrized.

Add QuicWriteBlockedListTest wrapper methods for all `write_blocked_list_`
public methods to make tests more concise.  Remove QuicStreamPriority type name
from argument calls to make tests even more concise.  These two together allow
repeated RegisterStream() calls to fit on a single line each, making tests
easier to read.

Add tests and augment existsing tests to cover previously not exercised methods:
GetSpdyPriorityOfStream(), UpdateStreamPriority(), UnregisterStream().

Remove references to crypto stream and headers stream in comments, as these only
exists in later versions of gQUIC, not in HTTP/3.  Use "static stream" instead,
which is not an HTTP/3 concept, but used widely in QUICHE source to refer to
streams that are essential (in the sense that closure is a connection error).

PiperOrigin-RevId: 502864670