Change PriorityWriteScheduler to use SpdyPriority.
PriorityWriteScheduler only uses the priority value of spdy::StreamPrecedence,
it ignores the parent stream and weight (if any).  For simplicity, this CL
changes RegisterStream() and UpdateStreamPriority() to take SpdyPriority
instead, and GetStreamPriority() to return one.

Also introduce PriorityWriteScheduler::PriorityType.  This is an alias for
spdy::SpdyPriority for now, and will become a template parameter in the next CL.

Also rename UpdateStreamPrecedence() to UpdateStreamPriority() and
GetStreamPrecedence() to GetStreamPriority() to reflect the argument type
change.  Also renames GetLatestEventWithPrecedence() to
GetLatestEventWithPriority(), both on PriorityWriteScheduler and
Http2Dispatcher, to reflect that it is only the priority (integer value) of the
stream that is considered in that method.

Also remove unused GetStreamChildren() method.

PiperOrigin-RevId: 502977611