Change variable types in BalsaHeaders::ParseTokenList().
QUICHE does not compile in Envoy on Windows after some recent C++20-related
fixes.  This motivated cl/504122725, but that did not fix the issue, because the
C++20-only (begin iterator, count) constructor signature is not available, and
an iterator cannot implicitly be converted to a pointer type on that platform.
The current change reverts cl/504122725 and instead uses
absl::string_view::data() which return absl::string_view::pointer_type instead
of iterator_type.
PiperOrigin-RevId: 505126369