Add a fallthrough case to QuicStreamPriority::Default()
Some Envoy compilations don't like this no-default switch statement, so add a QUICHE_BUG() to handle the case that no type matches.

Also add default copts to the quic_stream_priority build rules in an attempt to get QUICHE builds to match the Envoy failures, but it doesn't seem to make a difference.

PiperOrigin-RevId: 518929051