Updating PublicMetadataHashWithHKDF function according to the latest IETF draft before we use it in prod. Adding the latest test vectors as well.
This CL also makes sure that openssl RSA struct is not used as storage for the exponent augmented with the public metadata in blinder as well as the verifier in third_party/anonymous_tokens/cpp/crypto.

These changes will let us enable public metadata support in prod in follow-up CLs as well as accept empty public metadata as a valid value when public metadata support in enabled.

PiperOrigin-RevId: 518984326