Update fingerprint function to prevent intentional collisions
If the backend and clients don't agree on the fingerprint algo, verification will fail. We'll need to patch this into beryllium at the end of the workday friday so minimum testers are affected. No production traffic uses public metadata, so this isn't a reliability concern warranting a careful migration strategy.

PiperOrigin-RevId: 523406924