Add BalsaFrame functionality to invoke BalsaVisitorInterface::OnInterimHeaders().
This CL adds a BalsaFrame option set_use_interim_headers_callback() that
instructs BalsaFrame to invoke BalsaVisitorInterface::OnInterimHeaders() when
parsing interim headers. This will be helpful for delivering 103 Early Hints.

The eventual goal will be to default-enable the new callback and remove the
methods specific to 100 Continue in favor of this new functionality.

PiperOrigin-RevId: 524352366