Validate incoming ECN feedback.
Marks will not be sent in production until paired with a congestion control that reports that it supports ECT(1) via SendAlgorithmInterface.

The bulk of the logic in this CL validates that the received ECN feedback from the peer is valid according to RFC9000, and stops sending ECN if invalid.

Also stops sending ECN if it appears the path is dropping ECN packets.

Though not clear from this CL, the actual mechanism to write ECN is for the PacketWriter to read per_packet_options_. QuicConnection checks if all the conditions to write ECN are present and clears the bits if not. GFE changes to follow.

Protected by FLAGS_quic_reloadable_flag_quic_send_ect1.

PiperOrigin-RevId: 525846393