Add support for CreateContextForMultiPortPath running async.
This change modifies CreateContextForMultiPortPath to now take an std::function as an argument, and instead of returning an std::unique_ptr<QuicPathValidationContext> the unique_ptr is now passed into the std::function.

There is no behavioral change in how CreateContextForMultiPortPath or MaybeCreateMultiPortPath run.

This change is necessary for Chrome's implementation of CreateContextForMultiPortPath as it will need to be asynchronous. There is a WIP CL out for this change on the chrome side here: https://chromium-review.googlesource.com/c/chromium/src/+/4500184. This CL will need to be uploaded when these current changes are eventually merged into Chrome.

PiperOrigin-RevId: 529106195