Add a class to represent a sequence of BalsaHeaders.
This CL introduces BalsaHeadersSequence, which wraps a std::list<BalsaHeaders>
and an iterator. Users add BalsaHeaders to the BalsaHeadersSequence (owning),
and can then get a pointer (non-owning) to successive BalsaHeaders.

The immediate usage of BalsaHeadersSequence will be in SimpleClient
(RequestsTestHelper ecosystem, cl/527686928), but the long-term envisioned use
case will be in JetstreamSession, e.g., adding to BalsaHeadersSequence in the
readers and consuming the BalsaHeaders while driving the response pipeline [1].

[1] Roughly
http://google3/gfe/gfe2/jetstream/jetstream_session.cc;l=6760;rcl=528925001 and
http://google3/gfe/gfe2/jetstream/net_http_requester.cc;l=2560;rcl=528925001.

PiperOrigin-RevId: 529155933