Avoid interim response handling for 101 Switching Protocols responses in BalsaFrame.
When updating SimpleHttpClient to set_use_interim_headers_callback(true) in
cl/527686928, I found that the client would not receive WebSocket payloads as
body after receiving the 101 response in some cases. This was likely because of
the state resetting in [1]. The resetting is consistent with 100 Continue
behavior and will likely also be helpful with 103 Early Hints (and other
interim responses), but is not helpful for Upgrade/101 Switching Protocols.

This CL special-cases 101 Switching Protocols, excluding these responses from
other interim response handling (involving OnInterimHeaders()), because 101
represents the "final" headers on the current protocol.

The added unit test fails without this change (missing, headers callback aside,
the body callbacks. Balsa instead consumes the entire message but only outputs
the header portion. Perhaps the body was erroneously considered as the start of
"final" headers).

[1] http://google3/third_party/quiche/balsa/balsa_frame.cc;l=857;rcl=524360383

PiperOrigin-RevId: 529163242