Add quiche::HttpValidationPolicy::validate_transfer_encoding.
This should always be true for GFE2 so that behavior is not changed. However,
Envoy has its own Transfer-Encoding validation (currently in
source/common/http/http1/codec_impl.cc, which will be removed in favor of UHV),
so it is desirable to skip validation in Balsa when used in Envoy.

Adding Content-Length to test cases to avoid MAYBE_BODY_BUT_NO_CONTENT_LENGTH
warning.

PiperOrigin-RevId: 529654658