Fix compilation on Chrome OS
Chrome on Chrome OS is built with _FORTIFY_SOURCE=2 which causes there to be two definitions of recv() visible to the compiler. This means that the SyscallWrapper() function cannot decide which definition to use, and so cannot infer its template arguments.

Specify the template arguments explicitly to workaround this problem.

No functional changes.

PiperOrigin-RevId: 530360898