Log encrypted ClientHellos through QuicConnectionDebugVisitor
As TLS messages are encrypted, it can be useful to log the cleartext versions
for debugging. QUIC already has callbacks to observe CRYPTO frames, so we,
broadly, already do this.

However, when ECH is enabled, the ClientHelloInner is sent encrypted. For
debugging, BoringSSL exposes this through the message callback under a custom
"content type". Chromium, when using ECH over TCP, exposes this through NetLog.

This CL registers the same callback in QUICHE and exports it out of
QuicConnectionDebugVisitor. I went ahead and did both sending (client) and
receiving (server), though receiving is currently moot since we don't do ECH
for QUIC servers yet.

PiperOrigin-RevId: 532598203