When a QUIC client probes from a new socket via a new writer or respond to peer's probing on a new socket, check if the socket is write blocked.
Currently QuicConnetion blindly write packets to the new socket, and let the platform to decide how to handle socket blockage case. It is okay if the platform silently drop the probing packet in this case. But Chrome added CHECKs to disallow QUIC to write to socket which is already blocked: https://bugs.chromium.org/p/chromium/issues/detail?id=1376754#c10. So this change is to address the CHECK violation there and unblock their further debugging on the original issue.

PiperOrigin-RevId: 538886916