Update BalsaHeadersSequence and BalsaVisitorInterface to use std::unique_ptr for interim headers.
This CL updates the related methods BalsaHeadersSequence::Append() and
BalsaVisitorInterface::OnInterimHeaders() to use std::unique_ptr<BalsaHeaders>.
The motivation for this change is to allow users to retain valid references an
underlying BalsaHeaders even after transferring ownership to the
BalsaHeadersSequence (e.g., http://screen/4GbAo5Z7jsZxBCK from the
very-much-work-in-progress cl/536902857). If attempting to retain a reference
to a bare BalsaHeaders that is then std::move'd, the reference will not be
valid after the std::move (where the BalsaHeaders we want is now a
move-constructed BalsaHeaders elsewhere).
PiperOrigin-RevId: 540666570