Removes a DCHECK that can be reached by fuzzing.
HTTP request trailers are not a thing, and the fuzz test case that triggers this DCHECK involves bad client behavior, but this is the shortest path to resolving the test failure.

I managed to repro this in a test case: http://sponge2/839e289c-da7c-44c7-94ae-444ae8c40519

I'm not including the test case in this CL as I don't think it's a useful example, either in the positive or negative direction.

PiperOrigin-RevId: 547821099