Clear QuicSpdyStreamBodyManager in QuicSpdyStream::CloseReadSide().
Without this change, QuicSpdyStreamBodyManager may have fragments pointing to
freed buffers after QuicStream::CloseReadSide() is called, for example, when a
RST_STREAM frame arrives. QuicSpdyStream::Readv() should not be called in such a
situation, but if it is, that results in a use-after-free.

This bug was found by Clusterfuzz in Chrome.

Protected by FLAGS_quic_reloadable_flag_quic_clear_body_manager.

PiperOrigin-RevId: 549381785