Add an interface method for accepting trailers (with ownership) in Balsa.
This CL introduces BalsaVisitorInterface::OnTrailers() to deliver trailers with
ownership, with the eventual goal of replacing
BalsaVisitorInterface::ProcessTrailers() that delivers a non-owning reference
to trailers. The goal is to migrate [1] and move the trailers-processing logic
to a callback in BackendResponder, somewhat simplifying HttpReader (more
details in descendant cl/550634500). A stretch goal is to unconditionally
deliver trailers and then store them in HttpRequestInfo only if the conditions
in [2] hold true.

[1] http://google3/gfe/gfe2/jetstream/http_reader.cc;l=353;rcl=547269861
[2] http://google3/gfe/gfe2/jetstream/backend_responder.cc;l=1119-1120;rcl=549171119

PiperOrigin-RevId: 551249347