Use three bits of Connection ID to indicate config.
Connection IDs generated by LoadBalancerEncoder and QuicSiloIndexEncoder are never used in production, because we do not load any configs. Thus, the only part that needs to be flag-protected is the number of '1' bits added to the front of HASH_QUIC connection IDs (in CombinedConnectionIdGenerator).

Protected by FLAGS_quic_restart_flag_quic_3bit_config_id.

PiperOrigin-RevId: 554623294