Support integration of standalone AT library in Chrome
With these changes we should be able to move the BlindSignAuth implementation in Quiche to use the standalone Anonymous Tokens library when building in Chrome. In the meantime, this shouldn't break the existing BSA implementation that uses AT exported into Quiche.

For reference, how I plan to integrate this with the Chrome codebase can be found in this CL: https://chromium-review.googlesource.com/c/chromium/src/+/4549362

PiperOrigin-RevId: 566707025