Fix oghttp2 to handle the submission of trailers after sending data with END_STREAM.
This CL updates OgHttp2Session::WriteForStream() to gracefully handle the
scenario of having pending trailers to write after writing a DATA frame with
the END_STREAM flag, which locally half-closes the stream. This scenario can
arise if the user calls SubmitTrailers() for a stream that also has a
DataFrameSource with send_fin() true. Before, OgHttp2Session would infinite
loop. With this change, OgHttp2Session will instead close the stream with
INTERNAL_ERROR.

Note that the new oghttp2 behavior is still a functional diff from nghttp2
behavior, which appears to drop the data source (including pending body) when
trailers are submitted. OgHttp2Session prioritizes writing the body (and
trailers if submitted without signaling END_STREAM on the body) over writing
the trailers.

This change resolves a fuzzing bug with oghttp2.

This CL also adds trace logging for the sending of DATA frames.

PiperOrigin-RevId: 567320945