Always send the HTTP/3 setting indicating support for RFC 9297 (HTTP Datagrams)
Support for HTTP Datagrams was initially disabled by default because it was an evolving specification. Now that RFC 9297 has been published, we can simplify our codebase by default-enabling them. This matches what we did for the QUIC DATAGRAM transport parameter after RFC 9221 was published.

For use cases that need HTTP Datagram support (such as WebTransport), this change is a no-op because they were already overriding this configuration. For other use cases, the only behavior change is that we now send the setting. We will still silently ignore datagrams when received, as we did before.

This CL also minorly refactors a test to make it easier to debug.

Protected by FLAGS_quic_reloadable_flag_quic_enable_h3_datagrams.

PiperOrigin-RevId: 569300393