Use absl::InlinedVector for fragments in HttpHeaderBlock
quiche::HttpHeaderBlock::HeaderValue::fragments_ was a std::vector<> which meant
that it required an extra heap allocation for every header. Use an
absl::InlinedVector instead so that in the common case of only one value for a
header no extra heap allocation is needed.

This improves the performance of Chromium's CreateSpdyHeadersFromHttpRequest()
from 5424ns median to 4372ns median when measured using the benchmark at https://chromium-review.googlesource.com/c/chromium/src/+/4902677.

No functional change.

PiperOrigin-RevId: 570579948