Remove exception-throwing logic from QUICHE.
The current code is not exception-safe, and can be written more
efficiently in most cases.

Both {optional,StatusOr}::value and .at() perform bounds-checking
and will throw an exception (resulting in a SIGABRT and likely
process death for google3 binaries built with -fno-exceptions).

There are a few directories which are not included in this CL
(quic/moqt, quic/tools quiche/oblivious_http), as well as test-only
code.

PiperOrigin-RevId: 572643937