Fix UBSan issues in QUICHE
It is UB to pass `nullptr` to `memcpy` and friends, even when the byte count is zero. This is a language bug in C that got carried over to C++. UBSan flags these as issues, which is one of the blockers for running Chromium through UBSan.

`std::copy` and friends from `<algorithm>` do not have this issue, so use those instead.

PiperOrigin-RevId: 576257137