Do not serialize default `urgency` value, even if `incremental` is set.
Back in January it has been reported that a certain server implementation
incorrectly signals protocol error if they receive an HTTP/3 PRIORITY_UPDATE
frame with `incremental` set to true but `urgency` not serialized (indicating
default value). As a workaround, SerializePriorityFieldValue() was changed to
always serialize `urgency` when `incremental` is true in the following commit:
https://quiche.googlesource.com/quiche/+/566b33c9fa5b1723db04be3d40dcaf102344c323
(This is the same as cl/504634170.)

It has been confirmed that the server has been fixed in February, and the fix
has been widely deployed. This CL reverts the above commit.

See also https://crbug.com/1410007.

PiperOrigin-RevId: 584621524