QUIC Chat Client for very basic Media-over-QUIC transport testing.
The current state is that the client connects, exchanges SETUP, sends ANNOUNCE and gets ANNOUNCE_OK with an interop server, then SUBSCRIBEs to the catalog track and gets SUBSCRIBE_OK. It receives the catalog OBJECT and will issue SUBSCRIBEs for the tracks there.

If the server subscribes to the track the client announces, it will reply to that subscribe.

There is no capability to send OBJECT, or process non-catalog objects.

PiperOrigin-RevId: 588109420