Allow QUIC control frames to be sent while a packet is being processed.
Removes the restriction added in cr/574987377. Instead, changes the
behavior of QuicPacketCreator::SetMaxPacketLength() to allow it
to be called while frames are queued. In this case, the new max packet
length is applied to the *next* packet.

Note: uses a restart flag instead of a reloadable flag to avoid needing to
coordinate the change in QuicSession with the change in QuicPacketCreator.

Protected by FLAGS_quic_restart_flag_quic_allow_control_frames_while_procesing.

PiperOrigin-RevId: 588249854