Inline expectations into EXPECT_QUIC_BUGs
This prevents reading uninitialized results if EXPECT_QUIC_BUGs ever run
as death tests.

Chromium is in the process of removing the ability to cancel FATAL
LogMessages (incl. the underlying LOG(DFATAL)s that EXPECT_QUIC_BUGs are
detecting). With this change DefaultClientProofSource passes if
EXPECT_QUIC_BUG is defined as EXPECT_DCHECK_DEATH in Chromium.

PiperOrigin-RevId: 595468073