Copy the ECN codepoint in QuicReceivedPacket::Clone().
Cloned incoming packets will not carry a ECN codepoint even when the original packet has one, so the resulting ECN feedback will be incorrect -- this will cause the peer to abandon ECN, potentially impairing performance.

The relevant instance of cloned packets is Buffered Packets at the beginning of the connection. ECN peers will abandon ECN if early packets are buffered.

The only other instance is the QuicProxyConnectionMap. Fixing Clone() is necessary but not sufficient for getting QuicProxyDispatcher to forward ECN properly.

There are tests for both downstream effects of this bug.

Protected by FLAGS_quic_reloadable_flag_quic_clone_ecn.

PiperOrigin-RevId: 595547494