Fixes a code defect reported in https://github.com/envoyproxy/envoy/issues/31710.
In a very specific situation, OgHttp2Session can enter an infinite loop. This requires the following conditions:
* a stream has both response body data and trailers to write;
* the stream becomes blocked by stream-level flow control while writing the body; and
* the connection still has flow control quota available.

In this circumstance, OgHttp2Session::HasReadyStream() returns true because trailers_ready_ is not empty, but WriteForStream() cannot make progress.

The fix is to remove the stream ID from trailers_ready_ in GetNextReadyStream().

Protected by not protected, does not affect the GFE binary.

PiperOrigin-RevId: 597622882