Disable `QuicUdpSocketTest.ReadReceiveTimestamp` for android.
Before cl/583036215, the test "passes" because `__ANDROID_API__`==19 and the test early returns.

After cl/583036215, the test fails because `setsockopt(fd, SOL_SOCKET, SO_TIMESTAMPING, ..)` succeeds but future `recvmsg` on the same fd won't return a CMSG with type `SO_TIMESTAMPING`. It's unclear whether this is a Android emulator problem, or a Android problem.

I'm also unclear whether there are external users that depend on the android behavior(if this is a emulator-only problem), let me know your thoughts please.

PiperOrigin-RevId: 597645086