Disable Huffman encoding for QPACK headers sent over Hyperloop.
Add a DisableHuffmanEncoding() to QuicSpdySession which must be
called before Initialize(), which causes the QpackEncoder and
MetadataManager to be created with Huffman encoding disabled.

Both GfeQuicClientSession and GfeQuicBackendServerSession call
this method in their constructors, when the flag is true.

The actual behavior change is in QpackInstructionEncoder. The
rest of the changes are merely plumbing.

Protected by FLAGS_gfe2_reloadable_flag_disable_huffman_for_hyperloop.

PiperOrigin-RevId: 603097044