Fix issues required to build additional Quiche tests for Chrome
This fixes build issues and test failures that occur when attempting to build the following test files as part of Chrome builds:

 - common/btree_scheduler_test.cc
 - common/print_elements_test.cc
 - common/wire_serialization_test.cc
 - http2/test_tools/hpack_example_test.cc
 - quic/core/deterministic_connection_id_generator_test.cc
 - quic/core/tls_server_handshaker_test.cc
 - quic/core/quic_generic_session_test.cc
 - quic/tools/connect_tunnel_test.cc
 - quic/tools/connect_udp_tunnel_test.cc

Note that I've only tested this on Linux, so more minor changes might be needed to get all the tests passing on all platforms.

This CL doesn't introduce any meaningful changes in functionality.

PiperOrigin-RevId: 606606193