Fix issues with Hpack EXPECT_DEATH test on Android + ChromeOS
Certain build configurations of Chrome discard messages streamed to CHECK, so using EXPECT_DEATH with a regex won't always work. To account for this, this CL adds a new `EXPECT_QUICHE_DEATH` macro similar to the existing `EXPECT_QUCHE_DEBUG_DEATH` macro that we will override the implementation of in Chrome to use `EXPECT_CHECK_DEATH`.

Also, adds my name to allowed_authors in copy.bara.sky so this will be associated with me on Github. :)

PiperOrigin-RevId: 610533493