Add a new MetadataVisitor class to QuicSpdyStream. If registered, an instance of this new class will receive decoded HTTP/3 METADATA frames. If not, registered, HTTP/3 METADATA frames will be delivered as unknown HTTP/3 frames, which is current behavior.
This is motivated by implements HTTP/3 METADATA handling in Envoy. Without
this new visitor, we need to hook into the EnvoyQuicSession's OnUnknownFrame
methods. But this is really annoying because we receive frames on the Stream,
bounce them (at the QUICHE layer) over the the Session which parses them into
Headers and then need to bounce them back over to the Stream. This is awkward.
By using this new Visitor, the Stream itself can handle METADATA.

Protected by FLAGS_quic_reloadable_flag_quic_enable_http3_metadata_decoding.

PiperOrigin-RevId: 611542297