Move quiche_feature_flags to actual file in Piper.
This breaks the dependency on all the genrule implementation details.

I've added a new test that will fail whenever a new QUICHE flag is
added, and prompt the CL author to update this file.

The update command as provided in the METADATA check is:
```
$ blaze build gfe/gfe2:quiche_feature_flags_list && cp blaze-genfiles/gfe/gfe2/quiche_feature_flags_list.h third_party/quiche/common/
```

I've confirmed that this leads to a significant reduction in the size of the envoy build dependency graph:

```
$ wc -l deps_envoy*
  # cl/
  25581 deps_envoy_2024-02.txt
  38946 deps_envoy_less-quiche.txt
  25337 deps_envoy_more-quiche.txt
```

As such, this eliminates around 98% of the additional transitive
dependencies introduced by cl/605634099.

PiperOrigin-RevId: 612458693