Replace deprecated version of absl::HexStringToBytes
Replaces the deprecated version of absl::HexStringToBytes that returns std::string with the version that returns bool and populates a passed in string with the bytes. This CL is not expected to introduce any changes in functionality other than now verifying that absl::HexStringToBytes works successfully.

This allows the Chrome build of the library to build without the `-Wno-deprecated-declarations` clang flag.

PiperOrigin-RevId: 612568046