Retry Poll when system clock is faster than QuicClock
`poll` and `WSAPoll` use the system clock, which may run faster (skew from) the
`clock_` used in the event loop. In this case, no error is reported but the
event loop should retry until the timeout is met.

This is more prevalent on Windows (~1% of my tests) and especially in a VM
like Lexan (~33% of tests).

PiperOrigin-RevId: 614056093