Fix MoqtSessionTests with unitialized peer_role_.
This should never happen except in tests, because the SETUP always arrives to set the role. But this CL adds a default value, just in case future changes disrupt the checks around SETUP arriving first and containing Role.

Also fixed an old initialized value problem where the varint for the Role parameter isn't correctly formatted, but we don't exit.

PiperOrigin-RevId: 615050498