Add new metrics to measure throughput in QBONE.
Historically, QBONE relies on metrics exported by the GFE and QUIC to measure throughput in bytes and packets:

* `/gfe/gfe2/http/bytes_in` for throughput in bytes to prod.
* `/gfe/gfe2/http/bytes_out` for throughput in bytes from prod.
* `/gfe/gfe2/spdy4_over_quic/packets_quic_read` for throughput in packets to prod.
* `/gfe/gfe2/spdy4_over_quic/connection_packet_writer_total_packets_sent` for throughput in packets from prod.

The initial motivation for this CL was to add a `traffic_class` dimension to the above methods, but as that involved modified non-QBONE code I decided to implement new QBONE-specific metrics. Each of the metrics in this CL have two dimensions:

* `direction`: TO_PROD or FROM_PROD
* `traffic_class`: A string that represents the traffic class of the packets.

A follow up CL will change the dashboard to use these metrics instead of the old ones once this is fully rolled out.

PiperOrigin-RevId: 620055645