Add support for nested proxying in MASQUE client code
This bulk of this CL take the encapsulated MASQUE classes and turns them into subclasses of the underlying MASQUE classes. It then adds plumbing between these various classes so that all the constructors work correctly.

This CL also tweaks how the encapsulated max packet size is computed, by subtracting from the available room in the underlying connection.

Finally, this CL modifies the test tool to support a list of proxies separated by commas.

I've confirmed that this works for up to three proxies, and when we go to four the MTU becomes too small and things fail, as expected.

This code is only used for testing, so it's not flag-protected.

PiperOrigin-RevId: 627075681