Wire up xRPC client to talk to Phosphor
This CL introduces an xRPC (go/xrpc-in-gmscore-doc) client
to talk to Phosphor and connects it to the serivce Binder stub
effectively providing the end to end data flow of the module.
Currently only the getInitialData() method is added. Once this CL is in,
the same will be replicated for authAndSign().

Using xRPC is the recommended way to communicate with 1P services
and is cleaner than handrolling a GMSCoreNetworkEngine to send HTTP requests.
I verified with laskey@ that Phosphor supports direct GRPC requests.

A unit test has been added that spins up an in-process GRPC
fake server and currently runs a test against it.

Tested end to end with the real Phosphor server locally and
I was able to get the correct response.

#blindauth

PiperOrigin-RevId: 627759698