Make QuicIntervalDeque::Iterator random access.
cl/630952202 made it clear that QuicIntervalDeque::Iterator::operator+() and
operator+=() were never instantiated (since they would not have compiled). What
happened here is that given that QuicIntervalDeque::Iterator::iterator_category
was defined to be `forward_iterator_tag`, when std::lower_bound() called
std::advance(), the implementation repeatedly calling iterator++() was selected,
which has linear time completixy. In order to actually use
QuicIntervalDeque::Iterator::operator+=() and achieve constant time complexity
for std::advance(), this CL changes the iterator_category to
random_access_iterator_tag. It also makes operator+=() and operator-() public as
required. It also removes operator+(), which apparently is not necessary,
presubamly because it is either automatically generated by the standard library,
or not used at all. Finally, a DCHECK() is added as a counterpart of the
existing one, to guard against out-of-bounds in the opposite direction.
PiperOrigin-RevId: 631161498