Add some QuicIntervalDeque::Iterator methods.
GCC requires operator-=() and operator--() to be defined for std::advance for
random access iterators.

Also, `QUICHE_DCHECK_LE(0u, index_)` is always true as `index_` is unsigned,
causing a warning that is promoted to an error in some environments. This CL
verifies that the addition does not result in underflow before the it is
performed, not after.

PiperOrigin-RevId: 633296228